grunt.file.delete is not silent

+++ tasks/qunit_blanket_lcov.js
@@ -25,7 +25,9 @@ module.exports = function(grunt) {
         }
 
         grunt.file.mkdir(path.dirname(options.dest));
-        grunt.file.delete(options.dest);
+        if (grunt.file.exists(options.dest)) {
+            grunt.file.delete(options.dest);
+        }
 
         if (options.inject_reporter) {
             grunt.config.merge({
